mysite/requirements.txt
asgiref==3.2.3
backcall==0.1.0
colorama==0.4.3
decorator==4.4.2
Django==3.0.3
djangorestframework==3.11.0
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
parso==0.6.2
pickleshare==0.7.5
prompt-toolkit==3.0.3
Pygments==2.5.2
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.1.8
